rename Navbar to TopNavbar
