--- src/components/Navbar/index.ts
@@ -1,3 +1,0 @@
-import Navbar from './Navbar'
-export { NAVBAR_HEIGHT } from './Navbar.style'
-export default Navbar
