+++ src/components/TopNavbar/index.ts
@@ -1,0 +1,3 @@
+import TopNavbar from './TopNavbar'
+export { TOP_NAVBAR_HEIGHT } from './TopNavbar.style'
+export default TopNavbar
