+++ src/components/index.ts
@@ -2,7 +2,7 @@ export { default as LayoutWithRouting } from './LayoutWithRouting'
 export { default as VideoGallery } from './VideoGallery'
 export { default as CoverVideo } from './CoverVideo'
 export { default as ChannelGallery } from './ChannelGallery'
-export { default as Navbar } from './Navbar'
+export { default as TopNavbar } from './TopNavbar'
 export { default as VideoGrid } from './VideoGrid'
 export { default as PlaceholderVideoGrid } from './PlaceholderVideoGrid'
 export { default as VideoPreview } from './VideoPreviewWithNavigation'
