+++ src/views/BrowseView/BrowseView.style.ts
@@ -2,7 +2,7 @@ import styled from '@emotion/styled'
 import { CategoryPicker, InfiniteVideoGrid, Text } from '@/shared/components'
 
 import { colors, sizes, transitions, zIndex } from '@/shared/theme'
-import { NAVBAR_HEIGHT } from '@/components/Navbar'
+import { TOP_NAVBAR_HEIGHT } from '@/components/TopNavbar'
 
 type IsAtTop = {
   isAtTop: boolean
@@ -16,7 +16,7 @@ export const StyledCategoryPicker = styled(CategoryPicker)<IsAtTop>`
   z-index: ${zIndex.overlay};
   position: sticky;
   /*Offset Category Picker by Navbar Height */
-  top: ${NAVBAR_HEIGHT}px;
+  top: ${TOP_NAVBAR_HEIGHT}px;
   padding: ${sizes(5)} var(--global-horizontal-padding) ${sizes(2)};
   margin: 0 calc(-1 * var(--global-horizontal-padding));
   background-color: ${(props) => (props.isAtTop ? colors.transparent : colors.black)};
